Don't perform 64-bit accesses to the PLIC
Recent QEMU will fault for 8-byte accesses. Use a uint32_t instead of
uintptr_t to avoid those problems.

+++ machine/mtrap.h
@@ -43,9 +43,9 @@ typedef struct {
   volatile uint64_t* timecmp;
 
   volatile uint32_t* plic_m_thresh;
-  volatile uintptr_t* plic_m_ie;
+  volatile uint32_t* plic_m_ie;
   volatile uint32_t* plic_s_thresh;
-  volatile uintptr_t* plic_s_ie;
+  volatile uint32_t* plic_s_ie;
 } hls_t;
 
 #define MACHINE_STACK_TOP() ({ \
